+++ src/inspectors/quic5.c
@@ -35,6 +35,8 @@
 #include <openssl/bio.h>
 #include <openssl/evp.h>
 #include "quic_ssl_utils.h"
+#include "quic_tls13.h"
+#include "quic_utils.h"
 
 #define MAX_CONNECTION_ID_LENGTH 	20
 #define MAX_VERSION_LENGTH	 	4
@@ -115,124 +117,7 @@ static size_t convert_length_connection(size_t len){
   }
 }
 
-static size_t convert_length_sequence(size_t len){
-  switch(len){
-    case 0x30:
-      return 6;
-    case 0x20:
-      return 4;
-    case 0x10:
-      return 2;
-    case 0x00:
-      return 1;
-    default:
-      return 0;
-  }
-}
-
-static uint16_t quic_getu16(const unsigned char* start, size_t offset){
-#if __BYTE_ORDER == __LITTLE_ENDIAN
-  return get_u16((const char*) start, offset);
-#elif __BYTE_ORDER == __BIG_ENDIAN
-  uint16_t x = get_u16((const char*) start, offset);
-  return x << 8 | x >> 8;
-#else
-#error "Please fix <bits/endian.h>"
-#endif
-}
-
-static uint32_t quic_getu32(const unsigned char* start, size_t offset){
-#if __BYTE_ORDER == __LITTLE_ENDIAN
-  return get_u32((const char*) start, offset);
-#elif __BYTE_ORDER == __BIG_ENDIAN
-  uint32_t x = get_u32((const char*) start, offset);
-  return ((((x) & 0xff000000u) >> 24) | (((x) & 0x00ff0000u) >> 8) | (((x) & 0x0000ff00u) << 8) | (((x) & 0x000000ffu) << 24));
-#else
-#error "Please fix <bits/endian.h>"
-#endif
-}
-
-/* Quic variable length Integer decoding algorithm */
-size_t quic_get_variable_len(const unsigned char *app_data, size_t offset, uint64_t *var_len) {
-	size_t 		mbit	 	= app_data[offset] >> 6;
-	size_t 		len	 	= 0;
-
-	switch(mbit) {
-		case 0:
-			len = 1;
-			*var_len = (uint64_t)(app_data[offset] & 0x3F);
-			break;
-		case 1:
-			*var_len = ((uint64_t)(app_data[offset] & 0x3F) << 8) + (uint64_t)(app_data[offset + 1] & 0xFF);
-			len = 2;
-			break;
-		case 2:
-			*var_len = ((uint64_t)(app_data[offset] & 0x3F) << 24) + ((uint64_t)(app_data[offset + 1] & 0xFF) << 16)
-				 + ((uint64_t)(app_data[offset + 2] & 0xFF) << 8) + (uint64_t)(app_data[offset + 3] & 0xFF);
-			len = 4;
-			break;
-		case 3:
-			*var_len = ((uint64_t)(app_data[offset] & 0x3F) << 56) + ((uint64_t)(app_data[offset + 1] & 0xFF) << 48) 
-				+ ((uint64_t)(app_data[offset + 2] & 0xFF) << 40) + ((uint64_t)(app_data[offset + 3] & 0xFF) << 32)
-				+ ((uint64_t)(app_data[offset + 4] & 0xFF) << 24) + ((uint64_t)(app_data[offset + 5] & 0xFF) << 16) 
-				+ ((uint64_t)(app_data[offset + 6] & 0xFF) << 8) + (uint64_t)(app_data[offset + 7] & 0xFF);
-			len = 8;
-			break;
-		default:
-			len = 0; /* error should not happen */
-	}
-	return len;
-}
-
-/* START OF CODE THAT NEEDS TO BE REWRITTEN */
 #define MIN(X, Y) (((X) < (Y)) ? (X) : (Y))
-
-static void phton64(uint8_t *p, uint64_t v) {
-	p[0] = (uint8_t)(v >> 56);
-	p[1] = (uint8_t)(v >> 48);
-	p[2] = (uint8_t)(v >> 40);
-	p[3] = (uint8_t)(v >> 32);
-	p[4] = (uint8_t)(v >> 24);
-	p[5] = (uint8_t)(v >> 16);
-	p[6] = (uint8_t)(v >> 8);
-	p[7] = (uint8_t)(v >> 0);
-}
-
-static uint64_t pntoh64(const void *p) {
-	return (uint64_t)*((const uint8_t *)(p)+0)<<56|
-		(uint64_t)*((const uint8_t *)(p)+1)<<48|
-		(uint64_t)*((const uint8_t *)(p)+2)<<40|
-		(uint64_t)*((const uint8_t *)(p)+3)<<32|
-		(uint64_t)*((const uint8_t *)(p)+4)<<24|
-		(uint64_t)*((const uint8_t *)(p)+5)<<16|
-		(uint64_t)*((const uint8_t *)(p)+6)<<8|
-		(uint64_t)*((const uint8_t *)(p)+7)<<0;
-}
-
-static void debug_print_rawfield(const unsigned char *app_data, size_t start_offset, size_t len) {
-	size_t i;
-
-	for (i = 0; i < len; i++) {
-		printf("%02X", app_data[start_offset + i]);
-	}
-	printf("\n");
-}
-
-static void debug_print_charfield(const unsigned char *app_data, size_t start_offset, size_t len) {
-	size_t i;
-
-	for (i = 0; i < len; i++) {
-		printf("%C", app_data[start_offset + i]);
-	}
-	printf("\n");
-}
-
-static void *memdup(const uint8_t *orig, size_t len) {
-	void *dest = malloc(len);
-	if(dest)
-		memcpy(dest, orig, len);
-	return dest;
-}
 
 /**
  * Compute the client and server initial secrets given Connection ID "cid".
@@ -428,144 +313,6 @@ int decrypt_first_packet(quic_t *quic_info, const unsigned char *app_data, size_
 		return -1;
 	}
 	return 0;
-}
-
-void quic_parse_google_user_agent(pfwl_state_t *state, const unsigned char *data, size_t len, pfwl_dissection_info_t *pkt_info, pfwl_flow_info_private_t *flow_info_private) {
-        char *scratchpad = state->scratchpad + state->scratchpad_next_byte;
-        memcpy(scratchpad, data, len);
-        pfwl_field_string_set(pkt_info->l7.protocol_fields, PFWL_FIELDS_L7_QUIC_UAID, scratchpad, len);
-        state->scratchpad_next_byte += len;
-}
-
-void tls_parse_quic_transport_params(pfwl_state_t *state, const unsigned char *data, size_t len, pfwl_dissection_info_t *pkt_info, pfwl_flow_info_private_t *flow_info_private) {
-	size_t		pointer = 0;
-	size_t		TLVtype = 0;
-	size_t 		TLVlen 	= 0;
-	for (pointer = 0; pointer <len; pointer += TLVlen) {
-		pointer += quic_get_variable_len(data, pointer, &TLVtype);
-		TLVlen = data[pointer];
-		pointer++;
-		//printf("parameter TLV %08X TLV Size %02d\n", TLVtype, TLVlen);
-		switch(TLVtype) {
-			case 0x3129:
-				quic_parse_google_user_agent(state, data + pointer, TLVlen, pkt_info, flow_info_private);
-			default:
-				break;
-		}
-
-	}
-}
-
-void tls_parse_servername(pfwl_state_t *state, const unsigned char *data, size_t len, pfwl_dissection_info_t *pkt_info, pfwl_flow_info_private_t *flow_info_private) {
-	size_t		pointer = 0;
-	uint16_t 	list_len = ntohs(*(uint16_t *)(data));
-	size_t 		type 	= data[2];
-	uint16_t 	server_len = ntohs(*(uint16_t *)(data + 3));
-	pointer	= 2 + 1 + 2;
-
-	char *scratchpad = state->scratchpad + state->scratchpad_next_byte;
-	memcpy(scratchpad, data + pointer, server_len);
-	pfwl_field_string_set(pkt_info->l7.protocol_fields, PFWL_FIELDS_L7_QUIC_SNI, scratchpad, server_len);
-	state->scratchpad_next_byte += server_len;
-}
-
-void tls_parse_extensions(pfwl_state_t *state, const unsigned char *data, size_t len, pfwl_dissection_info_t *pkt_info, pfwl_flow_info_private_t *flow_info_private) {
-	size_t pointer;
-	size_t TLVlen;
-	size_t TLVtype;
-	for (pointer = 0; pointer < len; pointer += TLVlen) {
-		TLVtype = ntohs(*(uint16_t *)(&data[pointer]));
-		pointer += 2;
-		TLVlen = ntohs(*(uint16_t *)(&data[pointer]));
-		pointer += 2;
-		//printf("TLV %02d TLV Size %02d\n", TLVtype, TLVlen);
-		switch(TLVtype) {
-			/* Server Name */
-			case 0:
-				tls_parse_servername(state, data + pointer, TLVlen, pkt_info, flow_info_private);
-				break;
-				/* Extension quic transport parameters */
-			case 65445:
-				tls_parse_quic_transport_params(state, data + pointer, TLVlen, pkt_info, flow_info_private);
-				break;
-			default:
-				break;
-		}	
-
-	}
-}
-
-void check_tls13(pfwl_state_t *state, const unsigned char *tls_data, size_t tls_data_length, pfwl_dissection_info_t *pkt_info, pfwl_flow_info_private_t *flow_info_private) {
-        /* Finger printing */
-        unsigned char ja3_string[1024];
-        size_t ja3_string_len;
-
-	size_t 		tls_pointer	= 0;
-
-	/* Parse TLS record header */
-	size_t tls_record_frame_type = tls_data[tls_pointer];
-	tls_pointer++;
-
-	size_t tls_record_offset     = tls_data[tls_pointer];
-	tls_pointer++;
-
-	uint16_t tls_record_len	     = ntohs(*(uint16_t *)(&tls_data[tls_pointer]));
-	tls_pointer += 2;
-
-	/* Parse TLS Handshake protocol */
-	size_t	handshake_type = tls_data[tls_pointer];
-	tls_pointer++;
-
-	size_t 	length = (tls_data[tls_pointer] << 16) + (tls_data[tls_pointer+1] << 8) + tls_data[tls_pointer+2];
-	tls_pointer += 3;
-
-	uint16_t tls_version = ntohs(*(uint16_t *)(&tls_data[tls_pointer]));
-	tls_pointer += 2;
-
-	/* Build JA3 string */
-	ja3_string_len = sprintf(ja3_string, "%d,", tls_version);
-
-	if (handshake_type == 1) { /* We only inspect client hello which has a type equal to 1 */	
-		/* skipping random data 32 bytes */
-		tls_pointer += 32;
-
-		/* skipping legacy_session_id one byte */
-		tls_pointer += 1;
-
-		/* Cipher suites and length */
-		uint16_t cipher_suite_len = ntohs(*(uint16_t *)(&tls_data[tls_pointer]));
-		tls_pointer += 2;
-
-		/* use content of cipher suite for building the JA3 hash */
-		for (size_t i = 0; i < cipher_suite_len; i += 2) {
-			uint16_t cipher_suite = ntohs(*(uint16_t *)(tls_data + tls_pointer + i));
-			ja3_string_len += sprintf(ja3_string + ja3_string_len, "%d-", cipher_suite);
-		}
-		if (cipher_suite_len) {
-			ja3_string_len--; //remove last dash (-) from ja3_string
-		}
-		ja3_string_len += sprintf(ja3_string + ja3_string_len, ",");
-		tls_pointer += cipher_suite_len;
-
-		/* compression methods length */
-		size_t compression_methods_len = tls_data[tls_pointer];
-		tls_pointer++;
-
-		/* Skip compression methods */
-		tls_pointer += compression_methods_len;
-
-		/* Extension length */
-		uint16_t ext_len = ntohs(*(uint16_t *)(&tls_data[tls_pointer]));
-		tls_pointer += 2;
-
-		/* Add Extension length to the ja3 string */
-		ja3_string_len += sprintf(ja3_string + ja3_string_len, "%d,,", ext_len);
-
-		/* lets iterate over the exention list */
-		unsigned const char *ext_data = tls_data + tls_pointer;
-		tls_parse_extensions(state, ext_data, ext_len, pkt_info, flow_info_private);
-	}
-	printf("JA3 String %s\n", ja3_string);
 }
 
 uint8_t check_quic5(pfwl_state_t *state, const unsigned char *app_data,
