+++ src/inspectors/quic_utils.h
@@ -1,0 +1,42 @@
+/*
+ * quic_utils.h
+ *
+ * Some simple quic helper functions
+ *
+ * =========================================================================
+ * Copyright (c) 2016-2019 Daniele De Sensi (d.desensi.software@gmail.com)
+ * Copyright (c) 2020 SoftAtHome (david.cluytens@gmail.com)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ * =========================================================================
+ */
+
+#include <peafowl/inspectors/inspectors.h>
+#include <peafowl/peafowl.h>
+
+uint16_t quic_getu16(const unsigned char* start, size_t offset);
+uint32_t quic_getu32(const unsigned char* start, size_t offset);
+size_t quic_get_variable_len(const unsigned char *app_data, size_t offset, uint64_t *var_len);
+void phton64(uint8_t *p, uint64_t v);
+uint64_t pntoh64(const void *p);
+
+void debug_print_rawfield(const unsigned char *app_data, size_t start_offset, size_t len);
+void debug_print_charfield(const unsigned char *app_data, size_t start_offset, size_t len);
+void *memdup(const uint8_t *orig, size_t len);
+
